fix(2d): allow custom `HighlightStyle` with `all`-color. Rewrite cache to store color for every char
Fix the `LezerHighlighter`, so it also works when the `all`-color is set
on the `HighlightStyle` (This is required to set the color of  e.g.
Python colons). The tree-traversal is not required to determine the
bounds of token, because `highlightTree` already provides those in the
`from` and `to` arguments to the `putStyle` function. The cache was
replaced to use a `Uint8Array` to store a color-id for every character,
which greatly simplifies color lookup in comparison to the `Map` that
stores index ranges as strings.